refactor cslb transformation
