+++ dap-etl/transformation/src/main/scala/org/broadinstitute/monster/dap/cslb/CslbTransformations.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.monster.dap
+package org.broadinstitute.monster.dap.cslb
 
 import org.broadinstitute.monster.dap.common.RawRecord
 import org.broadinstitute.monster.dogaging.jadeschema.table.Cslb
