+++ dap-etl/transformation/src/test/scala/org/broadinstitute/monster/dap/CslbTransformationSpec.scala
@@ -1,6 +1,7 @@
 package org.broadinstitute.monster.dap
 
 import org.broadinstitute.monster.dap.common.RawRecord
+import org.broadinstitute.monster.dap.cslb.CslbTransformations
 
 import java.time.LocalDate
 import java.time.format.DateTimeParseException
